Made a start with switching to react
