+++ docs/index_react.html
@@ -1,0 +1,11 @@
+<!DOCTYPE html>
+<html>
+  <head>
+    <title>React and TypeScript</title>
+  </head>
+  <body>
+    <div id="root"></div>
+    <!-- Compiled TypeScript -->
+    <script src="dist/bundle.js" type="text/javascript"></script>
+  </body>
+</html>
