+++ package.json
@@ -1,0 +1,47 @@
+{
+  "name": "ed3d_map",
+  "version": "1.0.0",
+  "description": "3D map for Elite:Dangerous",
+  "main": "index.ts",
+  "directories": {
+    "doc": "docs"
+  },
+  "scripts": {
+    "build": "webpack",
+    "watch": "webpack -w",
+    "test": "jest",
+    "test:tsc": "tsc -p tsconfig.test.json -w"
+  },
+  "repository": {
+    "type": "git",
+    "url": "git+https://github.com/dolfandringa/elitedangerous_3d_mapping.git"
+  },
+  "author": "",
+  "license": "GPL-3.0-or-later",
+  "bugs": {
+    "url": "https://github.com/dolfandringa/elitedangerous_3d_mapping/issues"
+  },
+  "homepage": "https://github.com/dolfandringa/elitedangerous_3d_mapping#readme",
+  "dependencies": {
+    "react": "^16.13.1",
+    "react-dom": "^16.13.1",
+    "semantic-ui-css": "^2.4.1",
+    "semantic-ui-react": "^1.2.1",
+    "three": "^0.120.1",
+    "typescript": "^4.0.2"
+  },
+  "devDependencies": {
+    "@types/jest": "^26.0.12",
+    "@types/react": "^16.9.49",
+    "@types/react-dom": "^16.9.8",
+    "css-loader": "^4.2.2",
+    "file-loader": "^6.1.0",
+    "jest": "^26.4.2",
+    "sass-loader": "^10.0.1",
+    "style-loader": "^1.2.1",
+    "ts-loader": "^8.0.3",
+    "url-loader": "^4.1.0",
+    "webpack": "^4.44.1",
+    "webpack-cli": "^3.3.12"
+  }
+}
