+++ scripts/show_clusters.py
@@ -1,0 +1,5 @@
+import open3d as o3d
+pcd = o3d.io.read_point_cloud("docs/barnards_clusters.ply")
+o3d.visualization.draw_geometries(
+    [pcd],
+)
