+++ src/index.tsx
@@ -1,0 +1,9 @@
+import * as React from 'react';
+import { render } from 'react-dom';
+import App from './App';
+import "semantic-ui-css/semantic.min.css";
+
+render(
+    <App />,
+    document.getElementById('root')
+)
