+++ tsconfig.json
@@ -1,0 +1,19 @@
+{
+    "compilerOptions": {
+        "allowSyntheticDefaultImports": true,
+        "module": "ES2015",
+        
+        // Set React as the JSX factory
+        "jsx": "react",
+    
+        // Target ES6 for more modern browsers
+        "target": "es6",
+        "moduleResolution": "node",
+  
+        // Include typings from built-in lib declarations
+        "lib": ["es2015", "dom", "dom.iterable"],
+        
+        // Include module source maps for debugging
+        "sourceMap": true,
+    },
+}
