+++ tsconfig.test.json
@@ -1,0 +1,8 @@
+{
+    "extends": "./tsconfig.json",
+    "include": [
+        "src/**/*.ts",
+        "src/**/*.test.ts",
+        "src/**/*.test.tsx",
+    ],
+}
